+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Only fetch distinct cities
+- Use select boxes for all filters
 
 ## [0.17.1] - 2025-03-13